Modified `ClearRect.rect_offset` to use `u32`
Since, in vulkan docs, it says that the rect must be inside the render
area, its better to use `u32` so that we don't get negative values since
its not applicable here.

I was using `i32` because in vulkan docs its using `VkRect2D` inside
`VkClearRect` which is a general rectangle object. Since ours is only
specific to `ClearRect` its better to make it specific.

+++ vulkano/src/command_buffer/sys.rs
@@ -709,8 +709,8 @@ impl UnsafeCommandBufferBuilder {
                 Some(ash::vk::ClearRect {
                     rect: ash::vk::Rect2D {
                         offset: ash::vk::Offset2D {
-                            x: rect.rect_offset[0],
-                            y: rect.rect_offset[1],
+                            x: rect.rect_offset[0] as i32,
+                            y: rect.rect_offset[1] as i32,
                         },
                         extent: ash::vk::Extent2D {
                             width: rect.rect_extent[0],
